+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. This change
 - autocommand events for `plugin-host` - `NeovimJavaPrepare` and `NeovimJavaReady`
 - fully qualified names as defaults for command/autocommand and request/notification handlers
 - support for automatically mapped arguments for `@NeovimCommand` and `@NeovimAutocommand` annotated methods - using Jackson
+- support for automatically mapped arguments for handler methods using custom mapper provided to `NeovimHandlerManager`
 
 ## [0.4.1] - 2022-06-04